requirements.txt
backoff==2.2.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.0
watchfiles==0.21.0
websockets==12.0
gunicorn==21.2.0
httpx==0.27.0
elastic-transport==8.12.0
elasticsearch==8.12.0
